requirements.txt
sentence-transformers==4.1.0
langchain==0.3.26
pypdf==5.6.1
python-dotenv
langchain-pinecone==0.2.8
langchain-community==0.3.26
-e.
ipykernel
langchain-groq
streamlit